requirements.txt
functions-framework==3.1.0
google-cloud-firestore==2.16.1
requests==2.27.1
flask==2.1.1
werkzeug==2.1.1